refactor(cli): unify sub-args via shim methods (S-subcmd C4-4b)
C4-4b: KvModeArgs sub-args type alignment + 3 shim accessors + 13 call site migration.

Changes:
- `KvModeArgs::kv_kivi_bits` u32 → u8 (legacy `kivi_bits` 타입 매칭).
- Args에 3 shim 추가:
  - `effective_kivi_bits() -> u8`
  - `effective_kivi_residual_size() -> usize`
  - `effective_kv_offload_storage() -> String`
- generate.rs 13건 직접 legacy 필드 read를 shim 호출로 교체:
  - `args.kivi_bits` × 7 → `args.effective_kivi_bits()`
  - `args.kivi_residual_size` × 4 → `args.effective_kivi_residual_size()`
  - `args.kv_offload` × 2 → `args.effective_kv_offload_storage()` (Offload path local bind)
- Shim 의미: `kv_mode_args.kv_mode` 가 Kivi/Offload면 새 필드, 그 외 legacy 필드 fallback.

테스트:
- `engine/tests/spec/test_kv_mode_args.rs` +6 신규 (총 12/12 PASS).
- session::cli 52/52, session 모듈 52/52 PASS.
- build/fmt clean. Clippy lib 회귀 (opencl/mod.rs unsafe ptr 1건 + doc lazy continuation)는
  사전 존재 (backlog [P2]). 내 변경 라인은 clean.
- backend::opencl::* 21 lib test + spec 4 fail은 device-required (backlog [P3]).

C4-4a (`50792ec8`) 후속. C4-4c 이후 candidate: legacy field hide/제거 (C5/C6).

+++ engine/src/session/cli/kv_mode.rs
@@ -16,7 +16,7 @@ pub struct KvModeArgs {
 
     /// KIVI quantization bits (kv-mode=kivi 한정)
     #[arg(long = "kv-kivi-bits", default_value_t = 2)]
-    pub kv_kivi_bits: u32,
+    pub kv_kivi_bits: u8,
 
     /// KIVI residual buffer length (kv-mode=kivi 한정)
     #[arg(long = "kv-kivi-residual-len", default_value_t = 128)]
